fix: address review — move AdMob SPI extension to dedicated file
